refactor(workflows): extract useUnsavedChangesGuard from canvas-editor (phase 4)
Pulls the unsaved-changes navigation guard out of workflow-canvas-editor.tsx
into a generic hook: beforeunload, the TanStack history blocker, and the
capture-phase anchor-click interceptor (TanStack's blocker misses plain
`<a href>` navigation), plus the exit-confirmation dialog and a requestExit()
to drive from the cancel/back button.

The key-remount redesign turned out to be an unnecessary prerequisite. Rather
than delete the editor's props-change reset effect (which resets the guard's
confirmingExit/pendingNavigationHref alongside ~18 other fields), the hook
exposes `reset()` and the reset effect calls it — pure behavior-preserving
relocation, no page-wrapper change. The Save/Discard paths cross back via
`onSaveAndExit` (async→boolean: node-drawer-dirty guard, save, dirty reset) and
`onDiscardAndExit` callbacks, so completeExit + the dialog stay owned by the
hook. `isManagedSource`/`canSaveInPlace` move up a few lines so the hook can be
constructed before the reset effect references its `reset`.

The editor sheds the Dialog* + useTanStackRouter imports and the
RouteNavigationBlocker type. 1,306 -> 1,169.

The sibling copy-form hook is intentionally NOT taken as an in-file lift: its
handlers mutate ~8 editor-state fields and call buildDefinition/onSaved/onCopied
(an in-file hook would need ~12 injected deps), and its real payoff is the
cross-file dedup (the third copy in workflow-detail.tsx). Left for a deliberate
cross-file dedup; see tasks/plan.md.

Verified: typecheck + eslint + workflow-canvas-editor.test (29/0, incl. the
stubbed history.block flow) + workflows suite (420/0) + full suite (5124/0) +
binary build + isolated-home boot smoke (Workflows loads; client.js → 200).
